# login/docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'login'","description":"'login'","contact":{"name":"Daniel","email":"danielazo@Danielazo.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'login'"}],"paths":{"/players/":{"post":{"tags":["player-controller"],"summary":"Create a new Player","description":"Return a player created with your id. If a player exist with the userName or email, then return 404.Additionally, the mail must be valid and the password must have at least 8 characters and contains at least one number, one lowecase letter, one uppercas leter and one special character.","operationId":"savePlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Player"}}},"required":true},"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"400":{"description":"UserName or email already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/matches/":{"post":{"tags":["match-controller"],"operationId":"saveMatch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MatchDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Match"}}}}}}},"/login":{"post":{"tags":["login-controller"],"summary":"Login a player in the plataform","description":"Return the player logged in if the credentoals are correct","operationId":"loginPlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Credential"}}},"required":true},"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"404":{"description":"The credentials are invalid","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/login/v2":{"post":{"tags":["login-controller"],"summary":"Login a player in the plataform","description":"Return the player logged in if the credentoals are correct","operationId":"loginPlayer_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CredentialV2"}}},"required":true},"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"404":{"description":"The credentials are invalid","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/players/{id}":{"get":{"tags":["player-controller"],"summary":"Get a Player by id","description":"Return a player by the id. If the player doesn´t exist return 404","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/players/{id}/matches":{"get":{"tags":["player-controller"],"operationId":"getMatchesOfPlayer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Match"}}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Chaeck a healty of the app","description":"If the app it´s alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}}},"components":{"schemas":{"Player":{"required":["email","password","userName"],"type":"object","properties":{"id":{"title":"Player ID","type":"integer","description":"The player id","format":"int64","example":1},"userName":{"title":"Player user name","type":"string","description":"The Player user name","example":"myusername"},"password":{"title":"Player password","type":"string","description":"The password Player´s ","example":"Cordoba+1"},"email":{"title":"Player emaiñl","type":"string","description":"The email Player´s","example":"email@email.com"},"avatar":{"title":"Player avatar","type":"string","description":"The avatar Player´s","nullable":true,"example":"example:la url de la imagen..."},"lastLogin":{"title":"Player last login","type":"string","description":"The las login Player´s","format":"date-time","nullable":true}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"MatchDTO":{"required":["gameId","playerId"],"type":"object","properties":{"gameId":{"type":"integer","format":"int64"},"playerId":{"type":"integer","format":"int64"}}},"Game":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"code":{"type":"string"},"userName":{"type":"string"},"description":{"type":"string"},"rules":{"type":"string"}}},"Match":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"game":{"$ref":"#/components/schemas/Game"},"player":{"$ref":"#/components/schemas/Player"},"createdDate":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["STARTED","FINISHED","CANCELED"]}}},"Credential":{"type":"object","properties":{"identity":{"oneOf":[{"$ref":"#/components/schemas/EmailIdentity"},{"$ref":"#/components/schemas/UserNameIdentity"}]},"password":{"type":"string"}}},"EmailIdentity":{"required":["email","identity_type"],"type":"object","allOf":[{"$ref":"#/components/schemas/Identity"},{"type":"object","properties":{"email":{"title":"Email to legged in","type":"string","description":"The player email","example":"email@email.com"}}}]},"Identity":{"required":["identity_type"],"type":"object","properties":{"identity_type":{"title":"Type of identity used to legged in","type":"string","description":"the type of identity provides to logged in","example":"USERNAME or EMAIL","enum":["USERNAME","EMAIL"]}},"discriminator":{"propertyName":"identity_type"}},"UserNameIdentity":{"required":["identity_type","user_name"],"type":"object","allOf":[{"$ref":"#/components/schemas/Identity"},{"type":"object","properties":{"user_name":{"title":"Username","type":"string","description":"The player username","example":"myusername"}}}]},"CredentialV2":{"required":["identity","password"],"type":"object","properties":{"password":{"type":"string"},"identity":{"title":"Email or UserName to logged in","type":"string","description":"The player email or username","example":"email@email.com"}}}}}}